fo.3.0.0_FO-70 Refactor usePost and added expired token modal upon cookie expiration
