+++ src/modules/post/hooks/usePostsInputs.js
@@ -1,0 +1,18 @@
+import { useState } from "react";
+
+function usePostsInputs() {
+  const [title, setTitle] = useState("");
+  const [message, setMessage] = useState("");
+  const [id, setId] = useState(null);
+
+  return {
+    title,
+    message,
+    id,
+    setTitle,
+    setMessage,
+    setId,
+  };
+}
+
+export default usePostsInputs;
